Installed node.js
Installed node.js, and the respective components of such. Furthermore, created new virtual environments, one for a reflex app framework, and the other for usage in link with poetry. Am I making too many dependencies? Maybe. Is this going to be a grand experiment to probably fix many catastrophic errors and learn through trials, tribulations, and tears what makes a happy codebase amalgamation of horrors? Yes, and I'm looking forward to it with glee.